Changing the lat/lng in venue and price in event from numeric to float datatype this time

+++ backend/db/migrations/20240702014629-create-venue.js
@@ -36,11 +36,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       lng: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       createdAt: {
